+++ __project__/settings.py
@@ -110,7 +110,10 @@ USE_TZ = True
 # Static files (CSS, JavaScript, Images)
 # https://docs.djangoproject.com/en/4.1/howto/static-files/
 
-STATIC_URL = 'static/'
+STATIC_ROOT = env('STATIC_DIR')
+STATIC_URL = env('STATIC_URL')
+STATICFILES_DIRS = env('STATICFILES_DIRS')
+
 
 # Default primary key field type
 # https://docs.djangoproject.com/en/4.1/ref/settings/#default-auto-field